beepy/requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.7.4.post0
anyio==3.3.2
appdirs==1.4.4
astroid==2.7.2
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
blinker==1.4
bpyutils==0.2.0
brotlipy==0.7.0
bs4==0.0.1
cachelib==0.6.0
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
colorclass==2.2.0
comtypes==1.1.11
distlib==0.3.3
docopt==0.6.2
fast-youtube-search==0.0.8
ffmpeg==1.4
filelock==3.3.0
Flask==2.0.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gTTS==2.2.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
HTMLParser==0.0.2
httpcore==0.13.7
httpx==0.19.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.2
install==1.3.4
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lameenc==1.4.1
lazy-object-proxy==1.6.0
lxml==4.9.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mutagen==1.45.1
orjson==3.6.7
outcome==1.2.0
packaging==21.0
pafy==0.5.5
pep8==1.7.1
platformdirs==2.2.0
priority==2.0.0
psycopg2-binary==2.9.1
py-cord==2.1.3
py-cord-components==2.2.3
python-dotenv==0.21.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pydub==0.25.1
pylint==2.10.2
PyNaCl==1.4.0
pyparsing==2.4.7
pypl==40.10
PySocks==1.7.1
pyttsx3==2.90
quart==0.16.3
requests==2.26.0
rfc3986==1.5.0
rhvoice-wrapper==0.7.1
rhvoice-wrapper-data==0.4.1
SCons==4.3.0
selenium==4.4.3
setuptools==62.6.0
six==1.16.0
snap==0.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spotipy==2.19.0
SQLAlchemy==1.4.31
svgwrite==1.4.1
terminaltables==3.1.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
virtualenv==20.7.2
websockets==10.0
Werkzeug==2.0.2
wheel==0.37.1
wrapt==1.12.1
wsproto==1.0.0
youtube-dl==2021.6.6
youtube-search-python==1.4.7
yt-dlp==2021.9.2